# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SmartAI Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get document details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
  /documents/{id}/analyze:
    post:
      operationId: analyzeDocument
      tags: [documents]
      summary: Analyze a document with AI
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentAnalysisRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentAnalysisResult"
...
  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
...
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    Document:
      type: object
      required: [id, name, fileType, fileSize, createdAt, analyzed]
      properties:
        id:
          type: integer
        name:
          type: string
        fileType:
          type: string
        fileSize:
          type: integer
        createdAt:
          type: string
          format: date-time
        analyzed:
          type: boolean
        summary:
          type: ["string", "null"]
    DocumentAnalysisRequest:
      type: object
      required: [analysisType]
      properties:
        analysisType:
          type: string
          enum: [summary, keyPoints, qa, quiz]
    DocumentAnalysisResult:
      type: object
      required: [analysisType, result, documentId]
      properties:
        analysisType:
          type: string
        result:
          type: string
        documentId:
          type: integer
        qaItems:
          type: array
          items:
            $ref: "#/components/schemas/QAItem"
    QAItem:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
...
    UserProfile:
      type: object
      required: [id, displayName, preferredLanguage, theme, preferredModel, voiceEnabled]
      properties:
        id:
          type: string
        displayName:
          type: ["string", "null"]
        preferredLanguage:
          type: string
          enum: [en, hi, mr]
        theme:
          type: string
          enum: [light, dark, system]
        preferredModel:
          type: string
        voiceEnabled:
          type: boolean
        bio:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        preferredLanguage:
          type: string
          enum: [en, hi, mr]
        theme:
          type: string
          enum: [light, dark, system]
        preferredModel:
          type: string
        voiceEnabled:
          type: boolean
        bio:
          type: string
...