# lib/api-zod/src/generated/types/document.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartAI Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface Document {
  id: number;
  name: string;
  fileType: string;
  fileSize: number;
  createdAt: Date;
  analyzed: boolean;
  /** @nullable */
  summary?: string | null;
}