# lib/api-zod/src/generated/types/userProfileUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SmartAI Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileUpdatePreferredLanguage } from './userProfileUpdatePreferredLanguage';
import type { UserProfileUpdateTheme } from './userProfileUpdateTheme';

export interface UserProfileUpdate {
  displayName?: string;
  preferredLanguage?: UserProfileUpdatePreferredLanguage;
  theme?: UserProfileUpdateTheme;
  preferredModel?: string;
  voiceEnabled?: boolean;
  bio?: string;
}